src/requirements.txt
Django==4.2.15
djangorestframework==3.15.2
djangorestframework-simplejwt
gunicorn==23.0.0
psycopg2-binary==2.9.9